move model class into independent class
