+++ src/with_sample_data/honey_import/main.py
@@ -1,15 +1,5 @@
 from file_system_loader import load_xls_binary_data_from_file
-
-# contains honey importing data for an year and month
-class HoneyImportInfomation:
-    def __init__ (self, year, month, countries):
-        self.year = year
-        self.month = month
-        self.country_amount_map = {}
-        self.country_price_map = {}
-        for country in countries:
-            self.country_amount_map[country] = int(0)
-            self.country_price_map[country] = int(0)
+from honey_import_information import HoneyImportInfomation
 
 def parse_first_row(line):
     year = line[0]
